+++ berkeley-function-call-leaderboard/bfcl/constants/eval_config.py
@@ -61,6 +61,9 @@ UNDERSCORE_TO_DOT = [
     "yi-large-fc",
     "openbmb/MiniCPM3-4B-FC",
     "grok-beta",
+    "qwen/qwq-32b-FC-novita",
+    "meta-llama/llama-4-maverick-17b-128e-instruct-fp8-FC-novita",
+    "meta-llama/llama-4-scout-17b-16e-instruct-FC-novita",
 ]
 
 RED_FONT = "\033[91m"