+++ berkeley-function-call-leaderboard/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Berkeley Function Calling Leaderboard will be documented in this file.
 
+- [Apr 13, 2025] [#980](https://github.com/ShishirPatil/gorilla/pull/980): Integrate Novita AI as a third-party inference provider for the following open-source models:
+  - `Llama-4-Maverick-17B-128E-Instruct-FP8` (Prompt & FC)
+  - `Llama-4-Scout-17B-16E-Instruct` (Prompt & FC)
+  - `Qwen/QwQ-32B` (Prompt & FC),
 - [Apr 9, 2025] [#943](https://github.com/ShishirPatil/gorilla/pull/943): Retire the executable categories from the leaderboard. The following categories will be excluded from the evaluation pipeline:
   - `rest`
   - `exec_simple`